Update GoT for tm-load-test v0.6.2
Updates the relevant files/configuration to be able to support
`tm-load-test` v0.6.2 (the latest at this point in time). This updates
the `tm-load-test` configuration to work by way of command line
parameters instead of a configuration file, and takes the values of
these command line parameters from each experiment's `config.toml` file.